Major backend and windows printer refactoring
- Refactor windows printer approach to rely on `wmic`
- Move windows printer code to `printer.py`
- Move templates returning endpoints to `views/`
- Rename `data.py` to `database.py`
- Rename `database.py` to `middleware.py`
- Fix reject god permissions for `update` and `delete` user
- Move default endpoints from `__init__.py` to `main.py`
- Rename `ex_functions.py` to `utils.py`
- Remove `errorsh` blueprint and move `nojs` to `main.py`
- Move `version` to `constants.py`
- Fix operators offices selection for users `add` and `update`
- Move `get_with_alias.py` to `utils.py`
- Add exceptions logger to handle `500`
- Change the truthy and falsy numbers of `nojs` to `0,1`
