+++ app/helpers.py
@@ -1,8 +1,13 @@
+# -*- coding: utf-8 -*-
+''' This Source Code Form is subject to the terms of the Mozilla Public
+License, v. 2.0. If a copy of the MPL was not distributed with this
+file, You can obtain one at http://mozilla.org/MPL/2.0/. '''
+
 from functools import wraps
 from flask import current_app, flash, redirect, url_for
 from flask_login import current_user
 
-import app.data as data
+import app.database as data
 
 
 def is_god():
