+++ gt_cached.json
@@ -824,6 +824,13 @@
         "fr": "Erreur: Aucune t\u00e2che \u00e0 r\u00e9installer",
         "it": "Errore: nessuna attivit\u00e0 esistente da reimpostare"
     },
+    "Error: Office selected does not exist!": {
+        "ar": "\u062e\u0637\u0623: \u0645\u0643\u062a\u0628 \u0627\u0644\u0645\u062d\u062f\u062f \u0644\u0627 \u0648\u062c\u0648\u062f \u0644\u0647!",
+        "en": "Error: Office selected does not exist!",
+        "es": "Error: Oficina seleccionado no existe!",
+        "fr": "Erreur: Office n'existe pas!",
+        "it": "Errore: Ufficio selezionato non esiste!"
+    },
     "Error: failed to repeat announcement.": {
         "ar": "\u062e\u0637\u0623: \u0641\u0634\u0644 \u0641\u064a \u062a\u0643\u0631\u0627\u0631 \u0627\u0644\u0625\u0639\u0644\u0627\u0646.",
         "en": "Error: failed to repeat announcement.",
@@ -850,6 +857,9 @@
         "es": "Error: se requiere iniciar sesi\u00f3n para acceder a la p\u00e1gina",
         "fr": "Erreur: la connexion est requise pour acc\u00e9der \u00e0 la page",
         "it": "Errore: \u00e8 necessario il login per accedere alla pagina"
+    },
+    "Error: main admin account cannot be updated .": {
+        "en": "Error: main admin account cannot be updated ."
     },
     "Error: main admin account cannot be updated ..": {
         "ar": "\u062e\u0637\u0623: \u0644\u0627 \u064a\u0645\u0643\u0646 \u062a\u062d\u062f\u064a\u062b \u062d\u0633\u0627\u0628 \u0627\u0644\u0645\u0633\u0624\u0648\u0644 \u0627\u0644\u0631\u0626\u064a\u0633\u064a ..",
@@ -961,6 +971,7 @@
     },
     "Error: user name already exists, choose another name": {
         "ar": "\u062e\u0637\u0623: \u0627\u0633\u0645 \u0627\u0644\u0645\u0633\u062a\u062e\u062f\u0645 \u0645\u0648\u062c\u0648\u062f \u0628\u0627\u0644\u0641\u0639\u0644 \u060c \u0627\u062e\u062a\u0631 \u0627\u0633\u0645\u064b\u0627 \u0622\u062e\u0631",
+        "en": "Error: user name already exists, choose another name",
         "es": "Error: el nombre de usuario ya existe, elige otro nombre",
         "fr": "Erreur: le nom d'utilisateur existe d\u00e9j\u00e0, choisissez un autre nom",
         "it": "Errore: il nome utente esiste gi\u00e0, scegli un altro nome"
@@ -1005,6 +1016,7 @@
     },
     "Error: you must have available printer, to use printed": {
         "ar": "\u062e\u0637\u0623: \u064a\u062c\u0628 \u0623\u0646 \u064a\u0643\u0648\u0646 \u0644\u062f\u064a\u0643 \u0637\u0627\u0628\u0639\u0629 \u0645\u062a\u0648\u0641\u0631\u0629 \u060c \u0644\u0627\u0633\u062a\u062e\u062f\u0627\u0645 \u0627\u0644\u0637\u0628\u0627\u0639\u0629",
+        "en": "Error: you must have available printer, to use printed",
         "es": "Error: debe tener una impresora disponible, para usar impresa",
         "fr": "Erreur: vous devez disposer de l'imprimante disponible pour utiliser",
         "it": "Errore: \u00e8 necessario avere la stampante disponibile, da usare stampata"
@@ -1714,6 +1726,7 @@
     },
     "Notice: user is updated . ": {
         "ar": "\u0625\u0634\u0639\u0627\u0631: \u064a\u062a\u0645 \u062a\u062d\u062f\u064a\u062b \u0627\u0644\u0645\u0633\u062a\u062e\u062f\u0645.",
+        "en": "Notice: user is updated .",
         "es": "Aviso: el usuario est\u00e1 actualizado.",
         "fr": "Avis: l'utilisateur est mis \u00e0 jour.",
         "it": "Avviso: l'utente \u00e8 aggiornato."
@@ -3693,6 +3706,22 @@
         "es": "el \u00fanico atributo de actualizaci\u00f3n para la cuenta de administrador principal es contrase\u00f1a",
         "fr": "le seul attribut pouvant \u00eatre mis \u00e0 jour sur le compte administrateur principal est le mot de passe",
         "it": "l'unico attributo aggiornabile all'account amministratore principale \u00e8 la password"
+    },
+    "the role is disabled, it will be enabled once you add a new office": {
+        "ar": "\u0647\u0648 \u062f\u0648\u0631 \u0627\u0644\u0645\u0639\u0648\u0642\u064a\u0646\u060c \u0641\u0625\u0646\u0647 \u0633\u064a\u062a\u0645 \u062a\u0645\u0643\u064a\u0646 \u0628\u0645\u062c\u0631\u062f \u0625\u0636\u0627\u0641\u0629 \u0645\u0643\u062a\u0628\u0627 \u062c\u062f\u064a\u062f\u0627",
+        "en": "the role is disabled, it will be enabled once you add a new office",
+        "es": "el papel est\u00e1 desactivado, se activar\u00e1 una vez que se agrega una nueva oficina",
+        "it": "il ruolo \u00e8 disattivato, esso verr\u00e0 attivato una volta che si aggiunge un nuovo ufficio"
+    },
+    "the role is disabled, it will be enabled once you create an office": {
+        "en": "the role is disabled, it will be enabled once you create an office"
+    },
+    "the role is not enabled, it will be enabled once you add a new office": {
+        "ar": "\u0644\u0645 \u064a\u062a\u0645 \u062a\u0645\u0643\u064a\u0646 \u062f\u0648\u0631\u060c \u0641\u0625\u0646\u0647 \u0633\u064a\u062a\u0645 \u062a\u0645\u0643\u064a\u0646 \u0628\u0645\u062c\u0631\u062f \u0625\u0636\u0627\u0641\u0629 \u0645\u0643\u062a\u0628\u0627 \u062c\u062f\u064a\u062f\u0627",
+        "en": "the role is not enabled, it will be enabled once you add a new office",
+        "es": "el papel no est\u00e1 activado, s\u00f3lo se activar\u00e1 una vez que se agrega una nueva oficina",
+        "fr": "le r\u00f4le n'est pas activ\u00e9, il sera activ\u00e9 une fois que vous ajoutez un nouveau bureau",
+        "it": "il ruolo non \u00e8 abilitata, verr\u00e0 attivato una volta che si aggiunge un nuovo ufficio"
     },
     "title": {
         "ar": "\u0639\u0646\u0648\u0627\u0646",
