+++ requirements.txt
@@ -23,7 +23,7 @@ Flask-gTTS==0.12
 Flask-Less==0.5
 Flask-Login==0.4.0
 Flask-Mail==0.9.1
-Flask-Minify==0.11
+Flask-Minify==0.15
 Flask-Moment==0.9.0
 Flask-PageDown==0.2.2
 Flask-QRcode==1.0.1
