+++ templates/base.html
@@ -76,7 +76,7 @@
         <noscript>
             /* checking if nojs is on */
             <style>html{display:none;}</style>
-            <meta http-equiv="refresh" content="0.0;url=/nojs/3">
+            <meta http-equiv="refresh" content="0.0;url=/nojs/0">
         </noscript>
         <div class="widw">
             <center><img class="img-responsive mb-1"  style="width: 30%;" src="{{ url_for('static', filename='images/logo.png') }}" alt="Logo"></center>
@@ -222,7 +222,7 @@
                                 <ul class="dropdown-menu" role="menu">
                                     {% for curLang in languages %}
                                     <li {% if defLang == curLang[0] %} class="active" {% endif %}>
-                                        <a href="{{ url_for('lang_switch' , lang=curLang[0]) }}">
+                                        <a href="{{ url_for('language_switch' , language=curLang[0]) }}">
                                             <span class="fa fa-{% if curLang[0] == 'en' %}globe{% else %}language{% endif %}"></span>
                                             {{curLang[1]}}
                                         </a>
