+++ templates/base_s.html
@@ -36,7 +36,7 @@
     <body>
         <noscript>
             <style>html{display:none;}</style>
-            <meta http-equiv="refresh" content="0.0;url=/nojs/3">
+            <meta http-equiv="refresh" content="0.0;url=/nojs/0">
         </noscript>
         <div class="container-fluid">
             <div class="col-xs-12">
